pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
name = "notify_bot"
version = "0.1.0"
description = "Personal telegram Bot."
readme = "README.md"
authors = [{ name = "Daniel Ramirez", email = "dxas90@gmail.com" }]
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
  # Async SQLite
  "aiosqlite>=0.20",
  # Async HTTP client (replaces requests — avoids blocking the event loop)
  "httpx[http2]>=0.28",
  # Templating
  "jinja2>=3.1",
  # Telegram bot framework
  "python-telegram-bot[job-queue]==22.6",
  # Pinned transitive deps
  "anyio==4.12.1",
  "cachetools==7.0.1",
  "certifi==2025.11.12",
  "h11==0.16.0",
  "h2==4.3.0",
  "hpack==4.1.0",
  "httpcore==1.0.9",
  "hyperframe==6.1.0",
  "idna==3.11",
  "sniffio==1.3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "pytest-mock>=3.14",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "python-simplified-semver"
local_scheme = "no-local-version"
git_describe_command = ["git", "describe", "--dirty", "--tags", "--long", "--match", "v*"]

[tool.hatch.build.hooks.vcs]
version-file = "notify_bot/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["notify_bot"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/docs",
    "/tests",
    "/.git",
    "/.gitignore",
    "/.dockerignore",
    "/.env.example",
    "/Dockerfile",
    "/mise.toml",
    "/renovate.json5",
    "/uv.lock",
    "*.db",
    "*.db-*",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
